codec refactor - moved remaining implementations
