+++ elrond-codec/src/impl_for_types/impl_vec.rs
@@ -1,6 +1,135 @@
+use crate::codec_err::{DecodeError, EncodeError};
+use crate::nested_de::NestedDecode;
+use crate::nested_de_input::NestedDecodeInput;
+use crate::nested_ser::NestedEncode;
+use crate::nested_ser_output::NestedEncodeOutput;
+use crate::top_de::TopDecode;
+use crate::top_de_input::TopDecodeInput;
+use crate::top_ser::TopEncode;
+use crate::top_ser_output::TopEncodeOutput;
+use crate::{boxed_slice_into_vec, TypeInfo};
+use alloc::vec::Vec;
+
+impl<T: NestedEncode> TopEncode for Vec<T> {
+    #[inline]
+    fn top_encode<O: TopEncodeOutput>(&self, output: O) -> Result<(), EncodeError> {
+        self.as_slice().top_encode(output)
+    }
+
+    #[inline]
+    fn top_encode_or_exit<O: TopEncodeOutput, ExitCtx: Clone>(
+        &self,
+        output: O,
+        c: ExitCtx,
+        exit: fn(ExitCtx, EncodeError) -> !,
+    ) {
+        self.as_slice().top_encode_or_exit(output, c, exit);
+    }
+}
+
+impl<T: NestedDecode> TopDecode for Vec<T> {
+    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
+        if let TypeInfo::U8 = T::TYPE_INFO {
+            let bytes = input.into_boxed_slice_u8();
+            let bytes_vec = boxed_slice_into_vec(bytes);
+            let cast_vec: Vec<T> = unsafe { core::mem::transmute(bytes_vec) };
+            Ok(cast_vec)
+        } else {
+            let bytes = input.into_boxed_slice_u8();
+            let mut_slice = &mut &*bytes;
+            let mut result: Vec<T> = Vec::new();
+            while !mut_slice.is_empty() {
+                result.push(T::dep_decode(mut_slice)?);
+            }
+            Ok(result)
+        }
+    }
+
+    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
+        input: I,
+        c: ExitCtx,
+        exit: fn(ExitCtx, DecodeError) -> !,
+    ) -> Self {
+        if let TypeInfo::U8 = T::TYPE_INFO {
+            let bytes = input.into_boxed_slice_u8();
+            let bytes_vec = boxed_slice_into_vec(bytes);
+            let cast_vec: Vec<T> = unsafe { core::mem::transmute(bytes_vec) };
+            cast_vec
+        } else {
+            let bytes = input.into_boxed_slice_u8();
+            let mut_slice = &mut &*bytes;
+            let mut result: Vec<T> = Vec::new();
+            while !mut_slice.is_empty() {
+                result.push(T::dep_decode_or_exit(mut_slice, c.clone(), exit));
+            }
+            result
+        }
+    }
+}
+
+impl<T: NestedEncode> NestedEncode for Vec<T> {
+    #[inline]
+    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
+        self.as_slice().dep_encode(dest)
+    }
+
+    #[inline]
+    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
+        &self,
+        dest: &mut O,
+        c: ExitCtx,
+        exit: fn(ExitCtx, EncodeError) -> !,
+    ) {
+        self.as_slice().dep_encode_or_exit(dest, c, exit);
+    }
+}
+
+impl<T: NestedDecode> NestedDecode for Vec<T> {
+    fn dep_decode<I: NestedDecodeInput>(input: &mut I) -> Result<Self, DecodeError> {
+        let size = usize::dep_decode(input)?;
+        match T::TYPE_INFO {
+            TypeInfo::U8 => {
+                let bytes = input.read_slice(size)?;
+                let bytes_copy = bytes.to_vec(); // copy is needed because result might outlive input
+                let cast_vec: Vec<T> = unsafe { core::mem::transmute(bytes_copy) };
+                Ok(cast_vec)
+            },
+            _ => {
+                let mut result: Vec<T> = Vec::with_capacity(size);
+                for _ in 0..size {
+                    result.push(T::dep_decode(input)?);
+                }
+                Ok(result)
+            },
+        }
+    }
+
+    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
+        input: &mut I,
+        c: ExitCtx,
+        exit: fn(ExitCtx, DecodeError) -> !,
+    ) -> Self {
+        let size = usize::dep_decode_or_exit(input, c.clone(), exit);
+        match T::TYPE_INFO {
+            TypeInfo::U8 => {
+                let bytes = input.read_slice_or_exit(size, c, exit);
+                let bytes_copy = bytes.to_vec(); // copy is needed because result might outlive input
+                let cast_vec: Vec<T> = unsafe { core::mem::transmute(bytes_copy) };
+                cast_vec
+            },
+            _ => {
+                let mut result: Vec<T> = Vec::with_capacity(size);
+                for _ in 0..size {
+                    result.push(T::dep_decode_or_exit(input, c.clone(), exit));
+                }
+                result
+            },
+        }
+    }
+}
+
 #[cfg(test)]
 pub mod tests {
-
     use crate::test_util::check_top_encode_decode;
 
     #[test]
