+++ elrond-codec/src/impl_for_types/mod.rs
@@ -8,9 +8,13 @@ mod impl_empty;
 mod impl_encode_num_mimic;
 mod impl_encode_num_signed;
 mod impl_encode_num_unsigned;
+mod impl_non_zero_usize;
 mod impl_numbers;
 mod impl_option;
+mod impl_ref;
+mod impl_slice;
 mod impl_string;
 mod impl_top_encode_from_no_err;
 mod impl_tuple;
+mod impl_unit;
 mod impl_vec;
