+++ elrond-codec/src/nested_de.rs
@@ -1,8 +1,3 @@
-use alloc::boxed::Box;
-use alloc::string::String;
-use alloc::vec::Vec;
-use core::num::NonZeroUsize;
-
 use crate::codec_err::DecodeError;
 use crate::nested_de_input::NestedDecodeInput;
 use crate::TypeInfo;
@@ -60,21 +55,6 @@ pub fn dep_decode_from_byte_slice_or_exit<D: NestedDecode, ExitCtx: Clone>(
     result
 }
 
-impl NestedDecode for () {
-    const TYPE_INFO: TypeInfo = TypeInfo::Unit;
-
-    fn dep_decode<I: NestedDecodeInput>(_: &mut I) -> Result<(), DecodeError> {
-        Ok(())
-    }
-
-    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
-        _: &mut I,
-        _: ExitCtx,
-        _: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-    }
-}
-
 impl NestedDecode for u8 {
     const TYPE_INFO: TypeInfo = TypeInfo::U8;
 
@@ -88,169 +68,5 @@ impl NestedDecode for u8 {
         exit: fn(ExitCtx, DecodeError) -> !,
     ) -> Self {
         input.read_byte_or_exit(c, exit)
-    }
-}
-
-impl<T: NestedDecode> NestedDecode for Vec<T> {
-    fn dep_decode<I: NestedDecodeInput>(input: &mut I) -> Result<Self, DecodeError> {
-        let size = usize::dep_decode(input)?;
-        match T::TYPE_INFO {
-            TypeInfo::U8 => {
-                let bytes = input.read_slice(size)?;
-                let bytes_copy = bytes.to_vec(); // copy is needed because result might outlive input
-                let cast_vec: Vec<T> = unsafe { core::mem::transmute(bytes_copy) };
-                Ok(cast_vec)
-            },
-            _ => {
-                let mut result: Vec<T> = Vec::with_capacity(size);
-                for _ in 0..size {
-                    result.push(T::dep_decode(input)?);
-                }
-                Ok(result)
-            },
-        }
-    }
-
-    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
-        input: &mut I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        let size = usize::dep_decode_or_exit(input, c.clone(), exit);
-        match T::TYPE_INFO {
-            TypeInfo::U8 => {
-                let bytes = input.read_slice_or_exit(size, c, exit);
-                let bytes_copy = bytes.to_vec(); // copy is needed because result might outlive input
-                let cast_vec: Vec<T> = unsafe { core::mem::transmute(bytes_copy) };
-                cast_vec
-            },
-            _ => {
-                let mut result: Vec<T> = Vec::with_capacity(size);
-                for _ in 0..size {
-                    result.push(T::dep_decode_or_exit(input, c.clone(), exit));
-                }
-                result
-            },
-        }
-    }
-}
-
-impl NestedDecode for String {
-    fn dep_decode<I: NestedDecodeInput>(input: &mut I) -> Result<Self, DecodeError> {
-        let raw = Vec::<u8>::dep_decode(input)?;
-        match String::from_utf8(raw) {
-            Ok(s) => Ok(s),
-            Err(_) => Err(DecodeError::UTF8_DECODE_ERROR),
-        }
-    }
-
-    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
-        input: &mut I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        let raw = Vec::<u8>::dep_decode_or_exit(input, c.clone(), exit);
-        match String::from_utf8(raw) {
-            Ok(s) => s,
-            Err(_) => exit(c, DecodeError::UTF8_DECODE_ERROR),
-        }
-    }
-}
-
-impl NestedDecode for Box<str> {
-    #[inline]
-    fn dep_decode<I: NestedDecodeInput>(input: &mut I) -> Result<Self, DecodeError> {
-        Ok(String::dep_decode(input)?.into_boxed_str())
-    }
-
-    #[inline]
-    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
-        input: &mut I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        String::dep_decode_or_exit(input, c, exit).into_boxed_str()
-    }
-}
-
-impl NestedDecode for bool {
-    const TYPE_INFO: TypeInfo = TypeInfo::Bool;
-
-    fn dep_decode<I: NestedDecodeInput>(input: &mut I) -> Result<Self, DecodeError> {
-        match input.read_byte()? {
-            0 => Ok(false),
-            1 => Ok(true),
-            _ => Err(DecodeError::INVALID_VALUE),
-        }
-    }
-
-    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
-        input: &mut I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        match input.read_byte_or_exit(c.clone(), exit) {
-            0 => false,
-            1 => true,
-            _ => exit(c, DecodeError::INVALID_VALUE),
-        }
-    }
-}
-
-impl<T: NestedDecode> NestedDecode for Option<T> {
-    fn dep_decode<I: NestedDecodeInput>(input: &mut I) -> Result<Self, DecodeError> {
-        match input.read_byte()? {
-            0 => Ok(None),
-            1 => Ok(Some(T::dep_decode(input)?)),
-            _ => Err(DecodeError::INVALID_VALUE),
-        }
-    }
-
-    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
-        input: &mut I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        match input.read_byte_or_exit(c.clone(), exit) {
-            0 => None,
-            1 => Some(T::dep_decode_or_exit(input, c, exit)),
-            _ => exit(c, DecodeError::INVALID_VALUE),
-        }
-    }
-}
-
-impl<T: NestedDecode> NestedDecode for Box<T> {
-    fn dep_decode<I: NestedDecodeInput>(input: &mut I) -> Result<Self, DecodeError> {
-        Ok(Box::new(T::dep_decode(input)?))
-    }
-
-    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
-        input: &mut I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        Box::new(T::dep_decode_or_exit(input, c, exit))
-    }
-}
-
-impl NestedDecode for NonZeroUsize {
-    fn dep_decode<I: NestedDecodeInput>(input: &mut I) -> Result<Self, DecodeError> {
-        if let Some(nz) = NonZeroUsize::new(usize::dep_decode(input)?) {
-            Ok(nz)
-        } else {
-            Err(DecodeError::INVALID_VALUE)
-        }
-    }
-
-    fn dep_decode_or_exit<I: NestedDecodeInput, ExitCtx: Clone>(
-        input: &mut I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        if let Some(nz) = NonZeroUsize::new(usize::dep_decode_or_exit(input, c.clone(), exit)) {
-            nz
-        } else {
-            exit(c, DecodeError::INVALID_VALUE)
-        }
     }
 }
