+++ elrond-codec/src/nested_ser.rs
@@ -1,11 +1,7 @@
-use alloc::boxed::Box;
-use alloc::string::String;
-use alloc::vec::Vec;
-use core::num::NonZeroUsize;
-
 use crate::codec_err::EncodeError;
 use crate::nested_ser_output::NestedEncodeOutput;
 use crate::TypeInfo;
+use alloc::vec::Vec;
 
 /// Most types will be encoded without any possibility of error.
 /// The trait is used to provide these implementations.
@@ -51,244 +47,9 @@ pub fn dep_encode_to_vec<T: NestedEncode>(obj: &T) -> Result<Vec<u8>, EncodeErro
     Ok(bytes)
 }
 
-/// Adds the concantenated encoded contents of a slice to an output buffer,
-/// without serializing the slice length.
-/// Byte slice is treated separately, via direct transmute.
-pub fn dep_encode_slice_contents<T: NestedEncode, O: NestedEncodeOutput>(
-    slice: &[T],
-    dest: &mut O,
-) -> Result<(), EncodeError> {
-    match T::TYPE_INFO {
-        TypeInfo::U8 => {
-            // cast &[T] to &[u8]
-            let slice: &[u8] =
-                unsafe { core::slice::from_raw_parts(slice.as_ptr() as *const u8, slice.len()) };
-            dest.write(slice);
-        },
-        _ => {
-            for x in slice {
-                x.dep_encode(dest)?;
-            }
-        },
-    }
-    Ok(())
-}
-
-pub fn dep_encode_slice_contents_or_exit<T, O, ExitCtx>(
-    slice: &[T],
-    dest: &mut O,
-    c: ExitCtx,
-    exit: fn(ExitCtx, EncodeError) -> !,
-) where
-    T: NestedEncode,
-    O: NestedEncodeOutput,
-    ExitCtx: Clone,
-{
-    match T::TYPE_INFO {
-        TypeInfo::U8 => {
-            // cast &[T] to &[u8]
-            let slice: &[u8] =
-                unsafe { core::slice::from_raw_parts(slice.as_ptr() as *const u8, slice.len()) };
-            dest.write(slice);
-        },
-        _ => {
-            for x in slice {
-                x.dep_encode_or_exit(dest, c.clone(), exit);
-            }
-        },
-    }
-}
-
-impl NestedEncodeNoErr for () {
-    fn dep_encode_no_err<O: NestedEncodeOutput>(&self, _: &mut O) {}
-}
-
-impl<T: NestedEncode> NestedEncode for &[T] {
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        // push size
-        self.len().dep_encode(dest)?;
-        // actual data
-        dep_encode_slice_contents(self, dest)
-    }
-
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        // push size
-        self.len().dep_encode_or_exit(dest, c.clone(), exit);
-        // actual data
-        dep_encode_slice_contents_or_exit(self, dest, c, exit);
-    }
-}
-
-impl<T: NestedEncode> NestedEncode for &T {
-    #[inline]
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        (*self).dep_encode(dest)
-    }
-
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        (*self).dep_encode_or_exit(dest, c, exit);
-    }
-}
-
-impl NestedEncode for &str {
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        self.as_bytes().dep_encode(dest)
-    }
-
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        self.as_bytes().dep_encode_or_exit(dest, c, exit);
-    }
-}
-
-impl<T: NestedEncode> NestedEncode for Vec<T> {
-    #[inline]
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        self.as_slice().dep_encode(dest)
-    }
-
-    #[inline]
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        self.as_slice().dep_encode_or_exit(dest, c, exit);
-    }
-}
-
-impl NestedEncode for String {
-    #[inline]
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        self.as_bytes().dep_encode(dest)
-    }
-
-    #[inline]
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        self.as_bytes().dep_encode_or_exit(dest, c, exit);
-    }
-}
-
-impl NestedEncode for Box<str> {
-    #[inline]
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        self.as_ref().as_bytes().dep_encode(dest)
-    }
-
-    #[inline]
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        self.as_ref().as_bytes().dep_encode_or_exit(dest, c, exit);
-    }
-}
-
 // No reversing needed for u8, because it is a single byte.
 impl NestedEncodeNoErr for u8 {
     fn dep_encode_no_err<O: NestedEncodeOutput>(&self, dest: &mut O) {
         dest.push_byte(*self as u8);
-    }
-}
-
-impl<T: NestedEncode> NestedEncode for Option<T> {
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        match self {
-            Some(v) => {
-                dest.push_byte(1u8);
-                v.dep_encode(dest)
-            },
-            None => {
-                dest.push_byte(0u8);
-                Ok(())
-            },
-        }
-    }
-
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        match self {
-            Some(v) => {
-                dest.push_byte(1u8);
-                v.dep_encode_or_exit(dest, c, exit);
-            },
-            None => {
-                dest.push_byte(0u8);
-            },
-        }
-    }
-}
-
-impl<T: NestedEncode> NestedEncode for Box<T> {
-    #[inline(never)]
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        self.as_ref().dep_encode(dest)
-    }
-
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        self.as_ref().dep_encode_or_exit(dest, c, exit);
-    }
-}
-
-impl<T: NestedEncode> NestedEncode for Box<[T]> {
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        self.as_ref().dep_encode(dest)
-    }
-
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        self.as_ref().dep_encode_or_exit(dest, c, exit);
-    }
-}
-
-impl NestedEncode for NonZeroUsize {
-    #[inline]
-    fn dep_encode<O: NestedEncodeOutput>(&self, dest: &mut O) -> Result<(), EncodeError> {
-        self.get().dep_encode(dest)
-    }
-
-    #[inline]
-    fn dep_encode_or_exit<O: NestedEncodeOutput, ExitCtx: Clone>(
-        &self,
-        dest: &mut O,
-        c: ExitCtx,
-        exit: fn(ExitCtx, EncodeError) -> !,
-    ) {
-        self.get().dep_encode_or_exit(dest, c, exit);
     }
 }
