+++ elrond-codec/src/top_de.rs
@@ -1,12 +1,8 @@
 use alloc::boxed::Box;
-use alloc::string::String;
-use alloc::vec::Vec;
-use core::num::NonZeroUsize;
 
 use crate::codec_err::DecodeError;
 use crate::nested_de::*;
 use crate::top_de_input::TopDecodeInput;
-use crate::transmute::*;
 use crate::TypeInfo;
 
 /// Trait that allows zero-copy read of values from an underlying API in big endian format.
@@ -95,211 +91,4 @@ where
         exit(c, DecodeError::INPUT_TOO_LONG);
     }
     result
-}
-
-impl TopDecode for () {
-    const TYPE_INFO: TypeInfo = TypeInfo::Unit;
-
-    fn top_decode<I: TopDecodeInput>(_: I) -> Result<Self, DecodeError> {
-        Ok(())
-    }
-
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        _: I,
-        _: ExitCtx,
-        _: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-    }
-}
-
-impl<T: TopDecode> TopDecode for Box<T> {
-    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
-        T::top_decode_boxed(input)
-    }
-
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        input: I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        T::top_decode_boxed_or_exit(input, c, exit)
-    }
-}
-
-// Allowed to implement this because [T] cannot implement NestedDecode, being ?Sized.
-impl<T: NestedDecode> TopDecode for Box<[T]> {
-    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
-        if let TypeInfo::U8 = T::TYPE_INFO {
-            let bytes = input.into_boxed_slice_u8();
-            let cast_bytes: Box<[T]> = unsafe { core::mem::transmute(bytes) };
-            Ok(cast_bytes)
-        } else {
-            let vec = Vec::<T>::top_decode(input)?;
-            Ok(vec_into_boxed_slice(vec))
-        }
-    }
-
-    /// Quick exit for any of the contained types
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        input: I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        if let TypeInfo::U8 = T::TYPE_INFO {
-            let bytes = input.into_boxed_slice_u8();
-            let cast_bytes: Box<[T]> = unsafe { core::mem::transmute(bytes) };
-            cast_bytes
-        } else {
-            let vec = Vec::<T>::top_decode_or_exit(input, c, exit);
-            vec_into_boxed_slice(vec)
-        }
-    }
-}
-
-impl<T: NestedDecode> TopDecode for Vec<T> {
-    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
-        if let TypeInfo::U8 = T::TYPE_INFO {
-            let bytes = input.into_boxed_slice_u8();
-            let bytes_vec = boxed_slice_into_vec(bytes);
-            let cast_vec: Vec<T> = unsafe { core::mem::transmute(bytes_vec) };
-            Ok(cast_vec)
-        } else {
-            let bytes = input.into_boxed_slice_u8();
-            let mut_slice = &mut &*bytes;
-            let mut result: Vec<T> = Vec::new();
-            while !mut_slice.is_empty() {
-                result.push(T::dep_decode(mut_slice)?);
-            }
-            Ok(result)
-        }
-    }
-
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        input: I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        if let TypeInfo::U8 = T::TYPE_INFO {
-            let bytes = input.into_boxed_slice_u8();
-            let bytes_vec = boxed_slice_into_vec(bytes);
-            let cast_vec: Vec<T> = unsafe { core::mem::transmute(bytes_vec) };
-            cast_vec
-        } else {
-            let bytes = input.into_boxed_slice_u8();
-            let mut_slice = &mut &*bytes;
-            let mut result: Vec<T> = Vec::new();
-            while !mut_slice.is_empty() {
-                result.push(T::dep_decode_or_exit(mut_slice, c.clone(), exit));
-            }
-            result
-        }
-    }
-}
-
-impl TopDecode for String {
-    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
-        let raw = Vec::<u8>::top_decode(input)?;
-        match String::from_utf8(raw) {
-            Ok(s) => Ok(s),
-            Err(_) => Err(DecodeError::UTF8_DECODE_ERROR),
-        }
-    }
-
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        input: I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        let raw = Vec::<u8>::top_decode_or_exit(input, c.clone(), exit);
-        match String::from_utf8(raw) {
-            Ok(s) => s,
-            Err(_) => exit(c, DecodeError::UTF8_DECODE_ERROR),
-        }
-    }
-}
-
-impl TopDecode for Box<str> {
-    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
-        Ok(String::top_decode(input)?.into_boxed_str())
-    }
-
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        input: I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        String::top_decode_or_exit(input, c, exit).into_boxed_str()
-    }
-}
-
-impl TopDecode for bool {
-    const TYPE_INFO: TypeInfo = TypeInfo::Bool;
-
-    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
-        match input.into_u64() {
-            0 => Ok(false),
-            1 => Ok(true),
-            _ => Err(DecodeError::INPUT_OUT_OF_RANGE),
-        }
-    }
-
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        input: I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        match input.into_u64() {
-            0 => false,
-            1 => true,
-            _ => exit(c, DecodeError::INPUT_OUT_OF_RANGE),
-        }
-    }
-}
-
-impl<T: NestedDecode> TopDecode for Option<T> {
-    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
-        let bytes = input.into_boxed_slice_u8();
-        if bytes.is_empty() {
-            Ok(None)
-        } else {
-            let item = dep_decode_from_byte_slice::<T>(&bytes[1..])?;
-            Ok(Some(item))
-        }
-    }
-
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        input: I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        let bytes = input.into_boxed_slice_u8();
-        if bytes.is_empty() {
-            None
-        } else {
-            let item = dep_decode_from_byte_slice_or_exit(&bytes[1..], c, exit);
-            Some(item)
-        }
-    }
-}
-
-impl TopDecode for NonZeroUsize {
-    fn top_decode<I: TopDecodeInput>(input: I) -> Result<Self, DecodeError> {
-        if let Some(nz) = NonZeroUsize::new(usize::top_decode(input)?) {
-            Ok(nz)
-        } else {
-            Err(DecodeError::INVALID_VALUE)
-        }
-    }
-
-    fn top_decode_or_exit<I: TopDecodeInput, ExitCtx: Clone>(
-        input: I,
-        c: ExitCtx,
-        exit: fn(ExitCtx, DecodeError) -> !,
-    ) -> Self {
-        if let Some(nz) = NonZeroUsize::new(usize::top_decode_or_exit(input, c.clone(), exit)) {
-            nz
-        } else {
-            exit(c, DecodeError::INVALID_VALUE)
-        }
-    }
 }
